requirements.txt
langgraph
langchain
langchain-core
langchain-anthropic
python-dotenv
defusedxml
typing_extensions